+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "ContractsRevo/transaction-nft-contract",
 ]
 
+
 [workspace.dependencies]
 soroban-sdk = "22.0.6"
